Move geometry utils to compute module to avoid circular imports
