+++ desc/compute/__init__.py
@@ -41,6 +41,7 @@ from . import (
     _surface,
 )
 from .data_index import data_index
+from .geom_utils import rpz2xyz, rpz2xyz_vec, xyz2rpz, xyz2rpz_vec
 from .utils import (
     arg_order,
     compute,
