+++ desc/compute/_curve.py
@@ -1,7 +1,7 @@
 from desc.backend import jnp
-from desc.geometry.utils import rpz2xyz, rpz2xyz_vec, xyz2rpz, xyz2rpz_vec
 
 from .data_index import register_compute_fun
+from .geom_utils import rpz2xyz, rpz2xyz_vec, xyz2rpz, xyz2rpz_vec
 from .utils import cross, dot
