+++ desc/magnetic_fields.py
@@ -6,8 +6,8 @@ import numpy as np
 from netCDF4 import Dataset
 
 from desc.backend import jit, jnp, odeint
+from desc.compute import rpz2xyz_vec, xyz2rpz
 from desc.derivatives import Derivative
-from desc.geometry.utils import rpz2xyz_vec, xyz2rpz
 from desc.grid import Grid
 from desc.interpolate import _approx_df, interp2d, interp3d
 from desc.io import IOAble
